prefill: chunk the moe_vec z grid past the 65535 cap
moe_vec_q puts the routed-row count (tokens * top_k) on gridDim.z, which
the hardware caps at 65535 -- only gridDim.x reaches 2^31-1. At top_k = 6
the last prefill chunk that fitted was 10922 tokens (10923 * 6 = 65538),
so --max-prefill-length 16384 (z = 98304) failed the very first launch
with cudaErrorInvalidValue while 8192 (z = 49152) worked. Inherited from
the vendored llama.cpp/vLLM kernel; unrelated to our row_pitch_bytes
addition, which was already int64 throughout.

The row count has to stay on z. blockIdx.x is the fastest-varying axis
and consecutive x-blocks are consecutive weight rows of the SAME
(token, expert) pair, which keeps one expert's rows co-resident in L2 for
this bandwidth-bound GEMV; moving the count to x would scatter the weight
streaming. So slice z instead and pass each launch the base of its chunk
in a new z_offset argument -- the same shape as quantize_row_q8_1_cuda's
existing MAX_BLOCK_SIZE loop. An explicit offset rather than a bumped
base pointer, because the kernel derives token = z / topk from the
absolute z and a pointer scheme would also force every chunk stride to be
a multiple of topk.

Also promoted to 64-bit, so this does not just turn a hard crash into
silent corruption:
  - dst[z * nrows + row], which wrapped around T ~ 174762
  - token * token_stride
  - tokens/top_k in the launcher signatures, so the product cannot
    overflow at the call boundary
  - &x[off * kx] in quantize_row_q8_1_cuda, which wraps for off > 524288
    at kx = 4096 and genuinely sees large ky (= tokens * top_k on the
    down GEMV)

The 19 per-type launchers differed only in template arguments and all 19
had copied the capped launch, so they are now generated from one shared
moe_vec_launch helper via a macro instead of hand-maintained.

tests/kernels/test_moe_vec_large_rows.py covers 12288 and 16384 tokens at
top_k = 6 (z = 73728 / 98304) plus the top_k = 1 down-GEMV shape, and
asserts bit-identical results against the concatenation of sub-cap
slices. Verified sensitive: making the kernel ignore z_offset leaves the
"does not throw" cases green and fails all four value comparisons.

+++ python/freetoken/kernel/csrc/gguf/gguf_kernel.cu
@@ -67,7 +67,7 @@ static void quantize_row_q8_1_cuda(const scalar_t* x, void* vy, const int kx, co
     const dim3 num_blocks(block_num_x, num_blocks_y, 1);
     const dim3 block_size(CUDA_DEQUANTIZE_BLOCK_SIZE, 1, 1);
     quantize_q8_1<<<num_blocks, block_size, 0, stream>>>(
-        &x[off * kx], (int32_t*)vy + off * (kx_padded / 32 * 9), kx, kx_padded);
+        &x[(int64_t)off * (int64_t)kx], (int32_t*)vy + off * (kx_padded / 32 * 9), kx, kx_padded);
   }
 }
